Clarify "Noto fonts" vs. "the Noto fonts" usage
I once wrote "The Noto fonts" but Susanna preferred "Noto fonts". I ended up not writing this consistently. Of course we’d do "The" in singular "the Noto project" and "the Noto collection", because there is only one of that. But with Noto fonts vs. the Noto fonts, it’s a bit trickier. To me, "the Noto fonts" suggests the collection as a whole, while "Noto fonts" may suggest just some of these fonts, something you may choose out of the collection. Think of "iOS apps work on the iPhone" (any iOS app works on the iPhone) vs. "Tap Reset to remove the iOS apps" (all of them). "OpenType fonts require Unicode" (= any) vs. "The OpenType fonts take up 5 GB on my drive" (= all of them). That’s the principle I’d like to adopt. More specifically, I’d tend to use "Noto fonts" whenever possible, and reserve "the Noto fonts" for very specific cases when we really mean the entire collection. For example "the Noto fonts support all of Unicode 12" but "Noto fonts use Unicode".

I’ve made some changes in the texts in the Noto docs repo. A good example: "To use Noto fonts on your website, select the Noto fonts for the writing systems that your website text will be using." — To use Noto fonts = to use any of the Noto fonts. Select the Noto fonts = select the specific Noto fonts. In some cases, I’ve replaced "The Noto fonts" with "All Noto fonts" for even stronger clarification (...are licensed under the OFL).